+++ src/pages/ProductsPage/components/Profile/Profile.js
@@ -1,0 +1,10 @@
+
+const Profile = ({ product }) => {
+  <p>
+    {product?.id}
+    {product?.name}
+    {product?.category}
+  </p>
+}
+
+export default Profile;
